+++ Assets/Scripts/UnitSelectedVisual.cs
@@ -1,0 +1,37 @@
+using System;
+using UnityEngine;
+
+public class UnitSelectedVisual : MonoBehaviour
+{
+    [SerializeField] private Unit unit;
+
+    private MeshRenderer meshRenderer;
+
+    private void Awake()
+    {
+        meshRenderer = GetComponent<MeshRenderer>();
+    }
+
+    private void Start()
+    {
+        UnitActionSystem.Instance.OnSelectedUnitChanged += UnitActionSystem_OnSelectedUnitChanged;
+        UpdateVisual();
+    }
+
+    private void UnitActionSystem_OnSelectedUnitChanged(object sender, EventArgs empty)
+    {
+        UpdateVisual();
+    }
+
+    private void UpdateVisual()
+    {
+        if (unit == UnitActionSystem.Instance.GetSelectedUnit())
+        {
+            meshRenderer.enabled = true; // Show the visual if this unit is selected
+        }
+        else
+        {
+            meshRenderer.enabled = false; // Hide the visual if this unit is not selected
+        }
+    }
+}