+++ Dokumentaatio/Changelog.md
@@ -1,9 +1,11 @@
 # Changelog
 
-## [1.0.0] - 2024-03-26
+## [1.0.0] - 2024-04-11
 ### Lisätty
 - Verkkopelipohja toimii
 - Yksin ja Verkkopeli : Pelaaja voi liikuttaa yhtä yksikköä painamalla hiiren vasentanappia
+- The player have multiple units.
+- The player can select just one unit and move only that one.
 
 ## [1.1.0] - Tulossa 
 -